+++ qupath-core/src/main/java/qupath/lib/images/servers/ImageServers.java
@@ -100,6 +100,7 @@ public class ImageServers {
 			.registerSubtype(ReorderRGBServerBuilder.class, "swapRedBlue")
 			.registerSubtype(ColorDeconvolutionServerBuilder.class, "color_deconvolved")
 			.registerSubtype(NormalizedImageServerBuilder.class, "normalized")
+			.registerSubtype(TypeConvertImageServerBuilder.class, "typeConvert")
 			;
 
 	private static GsonTools.SubTypeAdapterFactory<BufferedImageNormalizer> normalizerFactory =
@@ -744,6 +745,37 @@ public class ImageServers {
 			if (newBuilder == builder)
 				return this;
 			return new NormalizedImageServerBuilder(getMetadata().orElse(null), newBuilder, normalizer);
+		}
+
+	}
+
+	static class TypeConvertImageServerBuilder extends AbstractServerBuilder<BufferedImage> {
+
+		private ServerBuilder<BufferedImage> builder;
+		private PixelType pixelType;
+
+		TypeConvertImageServerBuilder(ImageServerMetadata metadata, ServerBuilder<BufferedImage> builder, PixelType pixelType) {
+			super(metadata);
+			this.builder = builder;
+			this.pixelType = pixelType;
+		}
+
+		@Override
+		protected ImageServer<BufferedImage> buildOriginal() throws Exception {
+			return new TypeConvertImageServer(builder.build(), pixelType);
+		}
+
+		@Override
+		public Collection<URI> getURIs() {
+			return builder.getURIs();
+		}
+
+		@Override
+		public ServerBuilder<BufferedImage> updateURIs(Map<URI, URI> updateMap) {
+			ServerBuilder<BufferedImage> newBuilder = builder.updateURIs(updateMap);
+			if (newBuilder == builder)
+				return this;
+			return new TypeConvertImageServerBuilder(getMetadata().orElse(null), newBuilder, pixelType);
 		}
 
 	}