+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This is a *work in progress* for the next major release.
   * This can substantially speed up _some_ scripts that don't need to access the image data (e.g. for some measurement export)
 * Initial core support for stain normalization and background subtraction (https://github.com/qupath/qupath/pull/1554)
   * Experimental - not yet a full feature or available through the user interface!
+* Add `TransformedServerBuilder.convertType(PixelType)` to convert pixel types
 
 ### Bugs fixed
 * Tile export to .ome.tif can convert to 8-bit unnecessarily (https://github.com/qupath/qupath/issues/1494)